requirements.txt
openai>=1.0.0
pandas>=1.5.0
python-dotenv>=0.19.0 
PySide6>=6.6.0
PyInstaller>=6.0.0 
PyQt5>=5.15.10 